pyproject.toml
[project]
name = "kelp-core"
version = "0.0.3"
requires-python = ">=3.12"
description = "Metadata Toolkit for Databricks Spark and Declarative Pipelines"
dependencies = [
    "databricks-sdk>=0.80.0",
    "jinja2>=3.1.6",
    "pydantic<3",
    "pyyaml>=6.0",
    "typer>=0.23.0",
]
readme = "README.md"
authors = [
    {name = "BenSchr"}]

[project.urls]
Homepage = "https://github.com/benschr/kelp-core"
Documentation = "https://benschr.github.io/kelp-core/"
Repository = "https://github.com/benschr/kelp-core"

[project.scripts]
kelp = "kelp.cli.cli:main"

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "databricks-labs-dqx>=0.12.0",
    "pyspark[pipelines]>=4.1",
    "faker>=40.4.0",
    "pytest>=8.3.0",
    "pytest-mock>=3.14.0",
    "pytest-cov>=7.0.0",
]
docs = [
    "griffe-pydantic>=1.2.0",
    "griffe2md>=1.3.2",
    "mkdocs-material>=9.7.1",
    "mkdocstrings-python>=2.0.2",
]


[build-system]
requires = ["uv_build>=0.9.23,<0.10.0"]
build-backend = "uv_build"


[tool.uv.build-backend]
module-name = "kelp"


[tool.uv]
package = true

[tool.ruff]
line-length = 100
include = ["pyproject.toml", "src/**/*.py", "scripts/**/*.py", "tests/**/*.py"]
exclude = ["tests/fixtures/**"]

[tool.ruff.lint]
extend-select = [
    "B",
    "BLE",
    "S",
    "A",
    "C4",
    "I001",
    "T10",
    "ISC",
    "LOG",
    "G",
    "PIE",
    "T20",
    "PT", 
    "Q",
    "RSE",
    "RET",
    "SLF",
    "SIM",
    "TID",
    "TD",
    "PTH",
    "N",
    "PERF",
    "D419",
    "UP",
    "FURB",
    "RUF",
]
ignore=["RET504","SIM105"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101","SLF001"]
"src/kelp/cli/**/*.py" = ["T201"]
"scripts/**/*.py" = ["T201"]

[tool.ruff.lint.flake8-comprehensions]
allow-dict-calls-with-keyword-arguments = true

[tool.ruff.format]
docstring-code-format = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.ty.src]
include = ["src", "tests", "scripts"]
exclude = ["tests/fixtures"]
